Multiple Pokemon Per Subscription (#158)
* Add initial support for multiple Pokemon subscriptions

* Add migration and account for `All` keyword

* Uncommand subscription commands

* Fix issue with multi subs

* Fix issue with PvP Great+Ultra league subscriptions

* Rename typos

* Remove excess checks and rename variables again

* Update migration

* version bump

+++ src/Bot.cs
@@ -30,7 +30,6 @@
 
     // TODO: List all subscriptions with info command
     // TODO: IV wildcards
-    // TODO: Egg subscriptions (maybe)
 
     public class Bot
     {
@@ -1039,10 +1038,10 @@
                 if (pokemon == 0)
                     continue;
 
-                if (!MasterFile.Instance.Pokedex.ContainsKey((int)pokemon))
+                if (!MasterFile.Instance.Pokedex.ContainsKey(pokemon))
                     continue;
 
-                var pkmn = MasterFile.Instance.Pokedex[(int)pokemon];
+                var pkmn = MasterFile.Instance.Pokedex[pokemon];
                 var pkmnStats = stats[pokemon];
                 var chance = pkmnStats.Shiny == 0 || pkmnStats.Total == 0 ? 0 : Convert.ToInt32(pkmnStats.Total / pkmnStats.Shiny);
                 if (chance == 0)
